Homepage choice logic
